pyproject.toml
[project]
name = "omniweb-agent-engine"
version = "2.0.0"
description = "Omniweb AI Agent Engine — multi-tenant AI voice + text platform powered by ElevenLabs + Twilio"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Redis / job queue
    "redis[hiredis]>=5.0.0",
    "arq>=0.26.0",

    # Twilio (phone numbers + SMS)
    "twilio>=9.0.0",

    # OpenAI (post-call LLM processing)
    "openai>=1.50.0",

    # Stripe
    "stripe>=10.0.0",

    # HTTP client (ElevenLabs API calls)
    "httpx>=0.27.0",

    # Auth / security
    "PyJWT>=2.8.0",

    # Config
    "pydantic-settings>=2.5.0",
    "email-validator>=2.2.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "phonenumbers>=8.13.0",
    "pytz>=2024.1",
    "structlog>=24.0.0",
    "tenacity>=9.0.0",
    "orjson>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.32.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
